Invalid Names - Refactored Using Rename Variables Method
1. Description of the shortcoming (name, short description is possible): 

The code flaw specially on the name of variable `B_WIDTH’ and ‘B_HEIGHT’  is considered as Invalid Names, where both of the variable names resulting in confusion. 

2. Brief description of the selected refactoring method: 

What Method: 
To refactor the code, I used rename variable method.  

Why:
 Because the name of variable `B_WIDTH’ and ‘B_HEIGHT’ will confuse the reader about what is this variable used for. In will lead to misunderstanding when we read the code. Therefore, I changed it to become `SCREEN_WIDTH’ and `SCREEN_HEIGHT’. To give a clear names of the variables and what is this variable are used for.

3. Conclusion: 

By implementing the rename variable method, it will helps the reader to understand what is the variables used for. Because it has already clear names from both of the variables.