Conditional Complexity - Refactored Using Ternary Operator & Extract Method
1. Description of the shortcoming (name, short description is possible): 

The code flaw specially on `move()’ method is considered as conditional complexity. Since there are to many if on the method for determining the behavior based on certain conditions. 

2. Brief description of the selected refactoring method: 

What Method: 
To refactor the code, I used Ternary Operator & Extract method.  

Why:
 I used Ternary Operator to refactor the code because it provides a simple way to handle the conditional logic and it can simplified the code. By using Ternary Operator, it will allows us to express the direction of movement in a single line of code without need for multiple if conditions. After implementing Ternary Operator, the code will be more shorter since it extract the other if condition that complicate the method in to a simple code. By other words, we can say that we also implement Extract Method.

3. Conclusion: 

By implementing the Ternary Operator and also Extract Method, the code becomes more compact and in leads to improve the readability of codes. The refactored code simplifies the conditional complexity by replacing the multiple if conditions with a simple expression, resulting in cleaner code.