Invalid Names - Refactored Using Rename Variables Method
1. Description of the shortcoming (name, short description is possible): 

The code flaw specially on the name of variable `iid’, `iia’, `iih’, and `iib’ will leads to the confusion of the misunderstanding about what is the variable used for. 

2. Brief description of the selected refactoring method: 

What Method: 
To refactor the code, I used rename variable method.  

Why:
 Since the name of variable is very unusual.  I use rename variable method to choose the correct name that give clear explanation about what is the variable used for. Resulting to become `imageBall’,  `imageApple’, `imageHead’, and `imageBomb’.

3. Conclusion: 

By implementing the rename variable method, it will helps the reader to understand what is the variables used for. Because it has already clear names for the variables.