Conditional Complexity - Refactored Using Extract Method
1. Description of the shortcoming (name, short description is possible): 

The code flaw specially on `checkCollision()’ method is considered as conditional complexity. Since there are to many if on the method for determining the behavior based on certain conditions. 

2. Brief description of the selected refactoring method: 

What Method: 
To refactor the code, I used Extract method.  

Why:
 I used Extract Method to refactor the code because it provides a simple way to handle the conditional logic and it can simplified the code. From `checkCollision()’, I split up the code into multiple function based on their behavior that consists of `checkSnakeCollision()’, `wrapAroundScreen()’, and `stopGameIfLost()’. I extract those method to improve the understandability to the code to know what is this if condition used for.

3. Conclusion: 

By implementing Extract Code, the code become more understandable and readable. Since the original code is confusing, we don’t know what is this if condition related to. After we split the method, then the code become more clearer.